update access pass creation to delegate funding to serviceability (#32)
With this change in serviceability
https://github.com/malbeclabs/doublezero/pull/1335 we have moved the
funding of a new user account with rent and txn lamports to be an
instruction within the accesspass set transaction. This consolidates and
simplifies the logic performed by the sentinel to only require
validating the eligibility of the validator requesting access and
generating an access pass for their service key/client IP pair.