Post validator payment merkle roots (#42)
## Summary of Changes

This PR adds the ability to initialize and submit a transaction which
contains the total amount owned to validators, the number of validators,
and the merkle root for that epoch. It adds these steps to the worker so
this should be the basic e2e that is required for testing.


## Testing Verification
* A test that is currently ignored because it requires a running local
validator confirms tx results for initializing and submitting
transactions.