validator-debt: ensure distribution has passed calculation_allowed_timestamp (#93)
This PR adds a check to make sure that the distribution has been
initialized sufficiently before it tries to write debt for some epoch
using the `calculation_allowed_timestamp` in the distribution. It also
adds ` blockhash` field to `ComputedSolanaValidatorDebts` so it can be
compared against some point in time. Finally it renames the seed to
`solana_validator_debt` which is what will be used for testnet and
mainnet-beta.

Closes https://github.com/malbeclabs/doublezero/issues/1618 and
https://github.com/malbeclabs/doublezero/issues/1553
## Testing Evidence

This is a snippet of the derserialized record on the ledger showing the
blockhash addition:
```
ComputedSolanaValidatorDebts {
    blockhash: "9aYYGJ1UGFpMyBfKepJqpVr6cN9whM64PuAFmPfipLzn",
    epoch: 836,
    debts: [
        ComputedSolanaValidatorDebt {
            node_id: mrgn4sJJu5GBa5wbKyjuASzhyCifvcedGoLtpKjB3Wf,
            amount: 0,
        },
        ComputedSolanaValidatorDebt {
            node_id: popscoyTKVksa4TyTXw488b3vvFxM7qQEyTBeMQopKu,
            amount: 0,
        },
```

Unit tests have been updated to include the `blockhash` field and now
pass.